wip: switch to expression context as necessary
