+++ crates/mongodb-agent-common/src/query/mod.rs
@@ -1,4 +1,5 @@
 pub mod arguments;
+mod column_ref;
 mod constants;
 mod execute_query_request;
 mod foreach;
